run tests in docker image using multi-stage docker build (#71)
The motivation for this change is so that I can write Bitwarden integration tests and run them using the bw cli that is actually installed in the docker container. That feels necessary to me because a recent upgrade to the bw cli didn't work properly and I had to pin a specific version. I'd prefer it if the build errored out if the bw cli breaks something.